version 0.0.2

+++ Source/src/Utils.cs
@@ -26,18 +26,40 @@ namespace PythonShell.Utils {
             }
         }
 
-        private static async Task DownloadFile(String url, String destFile) {
+        private static async Task DownloadFile(String url, String destFile, String fileType = "text") {
             var client = new HttpClient();
             var resp = await client.GetAsync(url);
-            using (var fs = new FileStream(destFile, FileMode.CreateNew)) {
-                await resp.Content.CopyToAsync(fs);
+
+            if (fileType.ToLower() == "binary") {
+                File.WriteAllBytes(destFile, await resp.Content.ReadAsByteArrayAsync());
+            }
+            else if (fileType.ToLower() == "text") {
+                File.WriteAllText(destFile, await resp.Content.ReadAsStringAsync());
             }
         }
 
+        private static Process RunSettingCmd(String pythonToRun, String[] arguments) {
+            var process = new Process {
+                StartInfo = new ProcessStartInfo {
+                    FileName = pythonToRun,
+                    Arguments = String.Join(" ", arguments),
+                    UseShellExecute = false,
+                    RedirectStandardOutput = true, RedirectStandardError = true
+                }
+            }!;
+
+            process.Start();
+            process.WaitForExit();
+
+            return process;
+        }
+
         public static async Task InitializeApp(ShellConfig config, Boolean createDefaultEnv) {
+            Console.WriteLine("Initializing shell...");
             config.DefaultPythonVersion = Util.CheckPythonVersion(config.DefaultPythonVersion);
 
-            String userHomeDir = System.Environment.GetEnvironmentVariable("HOME")!;
+            Console.WriteLine("Check shell app directory...");
+            var userHomeDir = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
             config.AppDir = Path.Join(userHomeDir, ".python_shell.net");
             if (!Directory.Exists(config.AppDir)) {
                 Directory.CreateDirectory(config.AppDir);
@@ -50,39 +72,45 @@ namespace PythonShell.Utils {
             if (!Directory.Exists(config.InstanceDir)) {
                 Directory.CreateDirectory(config.InstanceDir);
             }
+            Console.WriteLine("Shell app directory check finished.");
 
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows)) {
+                Console.WriteLine("Check default python binary files...");
                 String pythonDir = Path.Join(config.AppDir, "python");
-                String pythonZipFile = Path.Join(config.TempDir, "python.zip");
-                await Util.DownloadFile($"https://www.python.org/ftp/python/${config.DefaultPythonVersion}/python-{config.DefaultPythonVersion}-embed-amd64.zip", pythonZipFile);
-                ZipFile.ExtractToDirectory(pythonZipFile, pythonDir);
-                File.Delete(pythonZipFile);
+                if (!Directory.Exists(pythonDir)) {
+                    String pythonZipFile = Path.Join(config.TempDir, "python.zip");
+                    await Util.DownloadFile($"https://www.python.org/ftp/python/{config.DefaultPythonVersion}/python-{config.DefaultPythonVersion}-embed-amd64.zip", pythonZipFile, "binary");
+                    ZipFile.ExtractToDirectory(pythonZipFile, pythonDir);
+                    File.Delete(pythonZipFile);
 
-                var pthFile = Util.GetPythonPthFile(config, pythonDir);
-                File.WriteAllText(
-                    pthFile,
-                    File.ReadAllText(pthFile).Replace("#import site", "import site"),
-                    System.Text.Encoding.UTF8
-                );
+                    var pthFile = Util.GetPythonPthFile(config, pythonDir);
+                    File.WriteAllText(
+                        pthFile,
+                        File.ReadAllText(pthFile).Replace("#import site", "import site")
+                    );
+                }
+                Console.WriteLine("Python check finished.");
 
-                config.DefaultPythonPath = Path.Join(pythonDir, "python");
+                config.DefaultPythonPath = Path.Join(pythonDir, "python.exe");
             }
 
             if (File.Exists(config.DefaultPythonPath)) {
-                var p = Process.Start(config.DefaultPythonPath, new String[] { "-m", "pip", "install", "pip", "--upgrade" })!;
-                p.WaitForExit();
+                Console.WriteLine("Default settings for virtualenv...");
+                var p = Util.RunSettingCmd(config.DefaultPythonPath, new String[] { "-m", "pip", "install", "pip", "--upgrade" });
                 if (p.ExitCode != 0) {
                     String pipInstallFile = Path.Join(config.TempDir, "get-pip.py");
-                    await Util.DownloadFile("https://bootstrap.pypa.io/pip/get-pip.py", pipInstallFile);
-                    Process.Start(config.DefaultPythonPath, new String[] { pipInstallFile }).WaitForExit();
+                    await Util.DownloadFile("https://bootstrap.pypa.io/pip/get-pip.py", pipInstallFile, "text");
+                    Util.RunSettingCmd(config.DefaultPythonPath, new String[] { pipInstallFile });
                     File.Delete(pipInstallFile);
                 }
 
-                Process.Start(config.DefaultPythonPath, new String[] { "-m", "pip", "install", "virtualenv", "--upgrade" }).WaitForExit();
+                Util.RunSettingCmd(config.DefaultPythonPath, new String[] { "-m", "pip", "install", "virtualenv", "--upgrade" });
+                Console.WriteLine("Virtualenv settings finished.");
             }
 
             String defaultEnvdir = Util.GetShellInstanceDir(config, "default");
             if (createDefaultEnv) {
+                Console.WriteLine("Creating default env...");
                 if (Directory.Exists(defaultEnvdir)) {
                     config.DefaultPythonEnvPath = Util.GetVirtualEnv(config, "default");
                 }
@@ -90,11 +118,14 @@ namespace PythonShell.Utils {
                     config.DefaultPythonEnvPath = Util.CreateShellInstance(config, "default").Python;
                 }
 
-                Util.InstallRequiresToEnv(config, config.DefaultPythonEnvPath);
+                // Util.InstallRequiresToEnv(config, config.DefaultPythonEnvPath);
+                Console.WriteLine("Default env created.");
             }
             // else if (File.Exists(config.DefaultPythonPath)) {
             //     Util.InstallRequiresToEnv(config, config.DefaultPythonPath);
             // }
+
+            Console.WriteLine("Shell initialized.");
         }
 
         public static String CheckPythonVersion(String rawPythonVersion) {
@@ -126,7 +157,7 @@ namespace PythonShell.Utils {
         }
 
         public static String GetPythonPthFile(ShellConfig config, String pythonDiir) {
-            return Path.Join(pythonDiir, $"python{config.DefaultPythonVersion.Replace($".{config.DefaultPythonVersion.Split(".").Last()}", "")}._pth");
+            return Path.Join(pythonDiir, $"python{config.DefaultPythonVersion.Replace($".{config.DefaultPythonVersion.Split(".").Last()}", "").Replace(".", "")}._pth");
         }
 
         public static void ClearShellInstance(ShellConfig config, String instanceName) {
@@ -140,27 +171,30 @@ namespace PythonShell.Utils {
             }
         }
 
-        public static InstanceMap CreateShellInstance(ShellConfig config, String? instanceName = null) {
+        public static InstanceMap CreateShellInstance(ShellConfig config, String? instanceName = null, Boolean echo = true) {
             instanceName = instanceName ?? DateTime.Now.ToString("yyyy.MM.dd.HH.mm.ss");
+            if (echo) Console.WriteLine($"Creating shell instance [{instanceName}]...");
             String instanceDir = Util.GetShellInstanceDir(config, instanceName);
             String envPython;
 
             if (!Directory.Exists(instanceDir)) {
                 Directory.CreateDirectory(instanceDir);
                 Directory.CreateDirectory(Path.Join(instanceDir, "temp"));
-                envPython = Util.CreateVirtualEnv(config, instanceName);
+                envPython = Util.CreateVirtualEnv(config, instanceName, echo: echo);
             }
             else {
                 Util.ClearShellInstance(config, instanceName);
                 envPython = Util.GetVirtualEnv(config, instanceName);
             }
+            if (echo) Console.WriteLine("Shell instance created.");
 
             return new InstanceMap {
                 Dir = instanceDir, Python = envPython
             };
         }
 
-        public static String CreateVirtualEnv(ShellConfig config, String instanceName, String[]? pythonRequires = null) {
+        public static String CreateVirtualEnv(ShellConfig config, String instanceName, String[]? pythonRequires = null, Boolean echo = true) {
+            if (echo) Console.WriteLine("Creating virtualenv...");
             pythonRequires = pythonRequires ?? config.PythonRequires;
 
             String envDir = Path.Join(Util.GetShellInstanceDir(config, instanceName), "env");
@@ -168,10 +202,11 @@ namespace PythonShell.Utils {
                 config.DefaultPythonEnvPath = Util.GetVirtualEnv(config, "default");
             }
 
-            Process.Start(config.DefaultPythonPath, new String[] { "-m", "virtualenv", envDir }).WaitForExit();
+            Util.RunSettingCmd(config.DefaultPythonPath, new String[] { "-m", "virtualenv", envDir });
             String envPython = Util.GetVirtualEnv(config, instanceName);
-            Process.Start(envPython, new String[] { "-m", "pip", "install", "pip", "--upgrade" }).WaitForExit();
+            Util.RunSettingCmd(envPython, new String[] { "-m", "pip", "install", "pip", "--upgrade" });
             Util.InstallRequiresToEnv(config, envPython, pythonRequires: pythonRequires);
+            if (echo) Console.WriteLine("Virtualenv created.");
 
             return envPython;
         }
@@ -180,7 +215,7 @@ namespace PythonShell.Utils {
             Directory.Delete(Util.GetShellInstanceDir(config, instanceName), true);
         }
 
-        public static InstanceMap GetShellInstance(ShellConfig config, String instanceName) {
+        public static InstanceMap GetShellInstance(ShellConfig config, String instanceName, Boolean echo = true) {
             String instanceDir = Util.GetShellInstanceDir(config, instanceName);
             if (Directory.Exists(instanceDir)) {
                 return new InstanceMap {
@@ -189,7 +224,7 @@ namespace PythonShell.Utils {
                 };
             }
             else {
-                return Util.CreateShellInstance(config, instanceName);
+                return Util.CreateShellInstance(config, instanceName, echo: echo);
             }
         }
 
@@ -204,18 +239,20 @@ namespace PythonShell.Utils {
             }
         }
 
-        public static void InstallRequiresToEnv(ShellConfig config, String envPython, String[]? pythonRequires = null) {
+        public static void InstallRequiresToEnv(ShellConfig config, String envPython, String[]? pythonRequires = null, Boolean echo = true) {
             pythonRequires = pythonRequires ?? config.PythonRequires;
 
+            if (echo) Console.WriteLine("Installing requirements...");
             if (pythonRequires == null && config.PythonRequireFile != null) {
-                Process.Start(envPython, new String[] { "-m", "pip", "install", "-r", config.PythonRequireFile }).WaitForExit();
+                Util.RunSettingCmd(envPython, new String[] { "-m", "pip", "install", "-r", config.PythonRequireFile });
             }
             else if (pythonRequires != null) {
                 String tempPythonRequireFile = Path.Join(config.TempDir, DateTime.Now.ToString("yyyy.MM.dd.HH.mm.ss"));
-                File.WriteAllText(tempPythonRequireFile, String.Join("\n", pythonRequires), System.Text.Encoding.UTF8);
-                Process.Start(envPython, new String[] { "-m", "pip", "install", "-r", tempPythonRequireFile }).WaitForExit();
+                File.WriteAllText(tempPythonRequireFile, String.Join("\n", pythonRequires));
+                Util.RunSettingCmd(envPython, new String[] { "-m", "pip", "install", "-r", tempPythonRequireFile });
                 File.Delete(tempPythonRequireFile);
             }
+            if (echo) Console.WriteLine("Requirements installed.");
         }
 
         public static void RemoveShellInstance(ShellConfig config, String instanceName) {
